requirements.txt
Flask==2.2.5
Flask-Cors==5.0.0
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-WTF==1.2.2
google-api-core==2.15.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.62.0
PyJWT==2.8.0
pymongo==4.9.2
requests==2.31.0
requests-oauthlib==1.3.1
textblob==0.18.0.post0
dropbox==12.0.2
